requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via discord-py
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via aiohttp
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via pynacl
charset-normalizer==3.4.5
    # via requests
davey==0.1.4
    # via discord-py
discord-py[voice]==2.7.1
    # via -r requirements.in
elevenlabs==2.39.0
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via elevenlabs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via elevenlabs
pydantic-core==2.41.5
    # via
    #   elevenlabs
    #   pydantic
pynacl==1.5.0
    # via discord-py
python-dotenv==1.2.2
    # via -r requirements.in
requests==2.32.5
    # via elevenlabs
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   elevenlabs
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==16.0
    # via elevenlabs
yarl==1.23.0
    # via aiohttp
